requirements.txt
python-twitter
redis